[chore][pkg/stanza] Cleanup tcp input operator files
