+++ pkg/stanza/operator/input/tcp/input_test.go
